将lmmp_sqr_移至mul.c；并删除sqr.c文件
