+++ src/lammp/mul.c
@@ -108,3 +108,16 @@ void lmmp_mul_n_(mp_ptr restrict dst, mp_srcptr restrict numa, mp_srcptr restric
     else
         lmmp_mul_fft_(dst, numa, n, numb, n);
 }
+
+void lmmp_sqr_(mp_ptr restrict dst, mp_srcptr restrict numa, mp_size_t na) {
+    if (na < MUL_TOOM22_THRESHOLD)
+        lmmp_sqr_basecase_(dst, numa, na);
+    else if (na < MUL_TOOM33_THRESHOLD)
+        lmmp_sqr_toom2_(dst, numa, na);
+    else if (na < MUL_TOOM44_THRESHOLD)
+        lmmp_sqr_toom3_(dst, numa, na);
+    else if (na < MUL_FFT_THRESHOLD)
+        lmmp_sqr_toom4_(dst, numa, na);
+    else
+        lmmp_mul_fft_(dst, numa, na, numa, na);
+}
